backend/requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
beautifulsoup4==4.13.4
black==25.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coverage==7.8.2
cryptography==44.0.3
cssselect==1.3.0
curl_cffi==0.11.1
DataRecorder==3.6.2
distro==1.9.0
DownloadKit==2.0.7
DrissionPage==4.1.0.18
et_xmlfile==2.0.0
exceptiongroup==1.3.0
Faker==37.3.0
filelock==3.18.0
flake8==7.2.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hydra-core==1.3.2
idna==3.10
iniconfig==2.1.0
isort==6.0.1
jiter==0.10.0
lxml==5.4.0
mccabe==0.7.0
msal==1.32.3
mypy==1.15.0
mypy_extensions==1.1.0
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.93.0
openpyxl==3.1.5
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.31.0
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyExecJS==1.5.1
pyflakes==3.3.2
PyJWT==2.10.1
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.3.5
pytest-cov==6.1.1
PyYAML==6.0.2
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.7
tenacity==8.5.0
tldextract==5.3.0
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
websocket-client==1.8.0
websockets==15.0.1
